+++ data/colors/vars/component_light.ts
@@ -1,12 +1,17 @@
 import {alpha, darken, get} from '../../../src/utils-v1'
 
 export default {
+  control: {
+    borderColor: {
+      emphasis: '#858F99'
+    }
+  },
   avatar: {
     bg: get('scale.white'),
     border: get('border.subtle'),
     stackFade: get('scale.gray.3'),
     stackFadeMore: get('scale.gray.2'),
-    childShadow: (theme: any) => `-2px -2px 0 ${alpha(get('scale.white'), 0.8)(theme)}`
+    childShadow: (theme: any) => `0 0 0 2px ${alpha(get('scale.white'), 0.8)(theme)}`
   },
   topicTag: {
     border: 'transparent'
@@ -43,7 +48,7 @@ export default {
     disabledBg: get('neutral.muted')
   },
   diffstat: {
-    additionBg: get('scale.green.4')
+    additionBg: get('success.emphasis')
   },
   timeline: {
     badgeBg: get('scale.gray.1') // needs to be opaque
@@ -83,11 +88,11 @@ export default {
 
     primary: {
       text: get('scale.white'),
-      bg: get('scale.green.4'),
+      bg: get('success.emphasis'),
       border: get('border.subtle'),
       shadow: (theme: any) => `0 1px 0 ${alpha(get('scale.black'), 0.1)(theme)}`,
       insetShadow: (theme: any) => `inset 0 1px 0 ${alpha(get('scale.white'), 0.03)(theme)}`,
-      hoverBg: '#2c974b',
+      hoverBg: get('scale.green.5'),
       hoverBorder: get('border.subtle'),
       selectedBg: darken(get('btn.primary.hoverBg'), 0.02),
       selectedShadow: (theme: any) => `inset 0 1px 0 ${alpha(get('scale.green.9'), 0.2)(theme)}`,
@@ -160,20 +165,32 @@ export default {
 
   switchTrack: {
     bg: get('scale.gray.1'),
-    border: get('scale.gray.3'),
+    hoverBg: darken(get('scale.gray.1'), 0.03),
+    activeBg: darken(get('scale.gray.1'), 0.05),
+    disabledBg: get('scale.gray.4'),
+    fg: get('fg.muted'),
+    disabledFg: get('fg.onEmphasis'),
+    border: 'transparent', // TODO: remove this in next major release
 
     checked: {
-      bg: get('scale.blue.0'),
-      hoverBg: get('scale.blue.1'),
-      activeBg: get('scale.blue.2'),
-      border: get('scale.blue.3'),
+      bg: get('accent.emphasis'),
+      hoverBg: '#0860CA', // accent-emphasis + scale-blue-7 @ 20% opacity
+      activeBg: '#0757BA', // accent-emphasis + scale-blue-7 @ 20% opacity
+      fg: get('fg.onEmphasis'),
+      disabledFg: get('fg.onEmphasis'),
+      border: 'transparent', // TODO: remove this in next major release
     }
   },
 
   switchKnob: {
+    bg: get('canvas.default'),
+    disabledBg: get('canvas.subtle'),
+    border: '#858F99', // control contrast border https://github.com/primer/primitives/pull/485
+
     checked: {
-      bg: get('scale.blue.5'),
-      disabledBg: get('scale.gray.5'),
+      bg: get('canvas.default'), // TODO: remove this in next major release
+      disabledBg: get('canvas.subtle'), // TODO: remove this in next major release
+      border: get('accent.emphasis'),
     }
   },
 
@@ -181,6 +198,7 @@ export default {
     bg: get('scale.gray.1'),
 
     button: {
+      bg: get('canvas.default'),
       hover: {
         bg: alpha(get('scale.gray.3'), 0.2),
       },
@@ -190,7 +208,7 @@ export default {
       },
 
       selected: {
-        border: get('scale.gray.5'),
+        border: get('scale.gray.4'),
       },
     },
   },