+++ src/filters/isColorWithAlpha.ts
@@ -2,7 +2,7 @@ import type StyleDictionary from 'style-dictionary'
 import {isColor} from './isColor'
 
 /**
- * @description Checks if token is source token
+ * @description Checks if token is color with an alpha value
  * @param arguments [StyleDictionary.TransformedToken](https://github.com/amzn/style-dictionary/blob/main/types/TransformedToken.d.ts)
  * @returns boolean
  */