+++ src/transformers/colorToHexAlpha.ts
@@ -13,5 +13,5 @@ export const colorToHexAlpha: StyleDictionary.Transform = {
   type: `value`,
   transitive: true,
   matcher: isColorWithAlpha,
-  transformer: (token: StyleDictionary.TransformedToken) => toHex(alpha(getTokenValue(token), token.alpha)),
+  transformer: (token: StyleDictionary.TransformedToken) => toHex(alpha(getTokenValue(token), token.alpha, token)),
 }